Noweb phaseout stage I: blib
glsim will no longer use noweb.  Noweb's literate programming sources
have been increasingly hard to maintain; as a result documentation has
been poor and out of date.  Plus, the format does not allow us to profit
from the source-code-editing tools that rely on source-code parsing
(such as Emacs Semantic package).

We have decided to switch to Doxygen for interface documentation.
Documentation of implementation details is not as clean with Doxygen,
but up-to-date interface documentation is vital; so priority has been
given to this.  Implementation will be documented by commenting the
sources; eventually separate implementation notes shall be written.
Doxygen documentation will serve as the main reference for the library,
with a tutorial/user guide planned in the future, in a format to be
decided.

With this commit noweb is no longer needed to build the library from the
maintainer sources.  Portions of documentation (in the doc directory)
and test programs are still in .nw files.