Reorganized NearestNeighbours interface and added max displacement computation
NearestNeighbours is now an abstract base class that serves as front
end to different implementations of nearest-neighbour search.
Neighbour structures are rebuilt on demand or when the accumulated
displacement reaches some value.

The implementation of the neighbour list is still naive (going through
all pairs), but since it is not rebuilt at every step (thanks to the
changes in MC and VVerletMD), there is a noticeable speed gain already
at 200 LJ particles (with cutoff at 2^(1/6) sigma and density=1.2).